Moving accessors to model
