+++ Jabberpoint/src/jabberPoint/JabberPoint.java
@@ -2,11 +2,11 @@ package jabberPoint;
 
 import javax.swing.JOptionPane;
 
+import jabberPoint.model.Accessor;
 import jabberPoint.model.Presentation;
 import jabberPoint.model.Style;
-import jabberPoint.view.Accessor;
+import jabberPoint.model.XMLAccessor;
 import jabberPoint.view.JabberpointFrame;
-import jabberPoint.view.XMLAccessor;
 
 import java.io.IOException;
