+++ Jabberpoint/src/jabberPoint/model/DemoPresentation.java
@@ -1,7 +1,5 @@
 package jabberPoint.model;
 
-import jabberPoint.view.Accessor;
-
 /** Een ingebouwde demo-presentatie
  * @author Ian F. Darwin, ian@darwinsys.com, Gert Florijn, Sylvia Stuurman
  * @version 1.1 2002/12/17 Gert Florijn
